Move AOT tests to AotTests, migrate missing regression test
