+++ src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/AotTests.cs
@@ -1,0 +1,342 @@
+using System;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Text.RegularExpressions;
+using Microsoft.Build.Framework;
+using Mono.Cecil;
+using NUnit.Framework;
+using Xamarin.ProjectTools;
+
+namespace Xamarin.Android.Build.Tests
+{
+	[Category ("Node-1"), Category ("AOT")]
+	[Parallelizable (ParallelScope.Children)]
+	public class AotTests : BaseTest
+	{
+		[Test, Category ("SmokeTests")]
+		public void BuildBasicApplicationReleaseProfiledAot ()
+		{
+			var proj = new XamarinAndroidApplicationProject () {
+				IsRelease = true,
+				AndroidEnableProfiledAot = true,
+			};
+			proj.SetProperty (proj.ActiveConfigurationProperties, "AndroidExtraAotOptions", "--verbose");
+			using (var b = CreateApkBuilder (Path.Combine ("temp", TestName))) {
+				Assert.IsTrue (b.Build (proj), "Build should have succeeded.");
+				StringAssertEx.ContainsRegex (@"\[aot-compiler stdout\] Using profile data file.*build.Xamarin.Android.startup\.aotprofile", b.LastBuildOutput, "Should use default AOT profile", RegexOptions.IgnoreCase);
+				StringAssertEx.ContainsRegex (@"\[aot-compiler stdout\] Method.*emitted at", b.LastBuildOutput, "Should contain verbose AOT compiler output", RegexOptions.IgnoreCase);
+			}
+		}
+
+		[Test]
+		public void BuildBasicApplicationReleaseProfiledAotWithoutDefaultProfile ()
+		{
+			var proj = new XamarinAndroidApplicationProject () {
+				IsRelease = true,
+				AndroidEnableProfiledAot = true,
+			};
+			proj.SetProperty (proj.ActiveConfigurationProperties, "AndroidUseDefaultAotProfile", "false");
+			using (var b = CreateApkBuilder (Path.Combine ("temp", TestName))) {
+				Assert.IsTrue (b.Build (proj), "Build should have succeeded.");
+				StringAssertEx.DoesNotContainRegex (@"\[aot-compiler stdout\] Using profile data file.*build.Xamarin.Android.startup.*\.aotprofile", b.LastBuildOutput, "Should not use default AOT profile", RegexOptions.IgnoreCase);
+			}
+		}
+
+
+		[Test]
+		[TestCaseSource (nameof (BuildTest.AotChecks))]
+		[Category ("SmokeTests")]
+		public void BuildAotApplicationAndÜmläüts (string supportedAbis, bool enableLLVM, bool expectedResult)
+		{
+			var path = Path.Combine ("temp", string.Format ("BuildAotApplication AndÜmläüts_{0}_{1}_{2}", supportedAbis, enableLLVM, expectedResult));
+			var proj = new XamarinAndroidApplicationProject () {
+				IsRelease = true,
+				BundleAssemblies = false,
+				AotAssemblies = true,
+			};
+			proj.SetProperty (KnownProperties.TargetFrameworkVersion, "v5.1");
+			proj.SetAndroidSupportedAbis (supportedAbis);
+			proj.SetProperty ("EnableLLVM", enableLLVM.ToString ());
+			bool checkMinLlvmPath = enableLLVM && (supportedAbis == "armeabi-v7a" || supportedAbis == "x86");
+			if (checkMinLlvmPath) {
+				// Set //uses-sdk/@android:minSdkVersion so that LLVM uses the right libc.so
+				proj.AndroidManifest = $@"<?xml version=""1.0"" encoding=""utf-8""?>
+<manifest xmlns:android=""http://schemas.android.com/apk/res/android"" android:versionCode=""1"" android:versionName=""1.0"" package=""{proj.PackageName}"">
+	<uses-sdk android:minSdkVersion=""{Xamarin.Android.Tools.XABuildConfig.NDKMinimumApiAvailable}"" />
+	<application android:label=""{proj.ProjectName}"">
+	</application>
+</manifest>";
+			}
+			using (var b = CreateApkBuilder (path)) {
+				if (!b.CrossCompilerAvailable (supportedAbis))
+					Assert.Ignore ($"Cross compiler for {supportedAbis} was not available");
+				if (!b.GetSupportedRuntimes ().Any (x => supportedAbis == x.Abi))
+					Assert.Ignore ($"Runtime for {supportedAbis} was not available.");
+				b.ThrowOnBuildFailure = false;
+				b.Verbosity = LoggerVerbosity.Diagnostic;
+				Assert.AreEqual (expectedResult, b.Build (proj), "Build should have {0}.", expectedResult ? "succeeded" : "failed");
+				if (!expectedResult)
+					return;
+				//NOTE: Windows has shortened paths such as: C:\Users\myuser\ANDROI~3\ndk\PLATFO~1\AN3971~1\arch-x86\usr\lib\libc.so
+				if (checkMinLlvmPath && !IsWindows) {
+					// LLVM passes a direct path to libc.so, and we need to use the libc.so
+					// which corresponds to the *minimum* SDK version specified in AndroidManifest.xml
+					// Since we overrode minSdkVersion=16, that means we should use libc.so from android-16.
+					StringAssertEx.ContainsRegex (@"\s*\[aot-compiler stdout].*android-16.arch-.*.usr.lib.libc\.so", b.LastBuildOutput, "AOT+LLVM should use libc.so from minSdkVersion!");
+				}
+				foreach (var abi in supportedAbis.Split (new char [] { ';' })) {
+					var libapp = Path.Combine (Root, b.ProjectDirectory, proj.IntermediateOutputPath,
+						"bundles", abi, "libmonodroid_bundle_app.so");
+					Assert.IsFalse (File.Exists (libapp), abi + " libmonodroid_bundle_app.so should not exist");
+					var assemblies = Path.Combine (Root, b.ProjectDirectory, proj.IntermediateOutputPath,
+						"aot", abi, "libaot-UnnamedProject.dll.so");
+					Assert.IsTrue (File.Exists (assemblies), "{0} libaot-UnnamedProject.dll.so does not exist", abi);
+					var apk = Path.Combine (Root, b.ProjectDirectory,
+						proj.IntermediateOutputPath, "android", "bin", "UnnamedProject.UnnamedProject.apk");
+					using (var zipFile = ZipHelper.OpenZip (apk)) {
+						Assert.IsNotNull (ZipHelper.ReadFileFromZip (zipFile,
+							string.Format ("lib/{0}/libaot-UnnamedProject.dll.so", abi)),
+							"lib/{0}/libaot-UnnamedProject.dll.so should be in the UnnamedProject.UnnamedProject.apk", abi);
+						Assert.IsNotNull (ZipHelper.ReadFileFromZip (zipFile,
+							"assemblies/UnnamedProject.dll"),
+							"UnnamedProject.dll should be in the UnnamedProject.UnnamedProject.apk");
+					}
+				}
+				Assert.AreEqual (expectedResult, b.Build (proj), "Second Build should have {0}.", expectedResult ? "succeeded" : "failed");
+				Assert.IsTrue (
+					b.Output.IsTargetSkipped ("_CompileJava"),
+					"the _CompileJava target should be skipped");
+				Assert.IsTrue (
+					b.Output.IsTargetSkipped ("_BuildApkEmbed"),
+					"the _BuildApkEmbed target should be skipped");
+			}
+		}
+
+		[Test]
+		[TestCaseSource (nameof (BuildTest.AotChecks))]
+		[Category ("Minor"), Category ("MkBundle")]
+		public void BuildAotApplicationAndBundleAndÜmläüts (string supportedAbis, bool enableLLVM, bool expectedResult)
+		{
+			var path = Path.Combine ("temp", string.Format ("BuildAotApplicationAndBundle AndÜmläüts_{0}_{1}_{2}", supportedAbis, enableLLVM, expectedResult));
+			var proj = new XamarinAndroidApplicationProject () {
+				IsRelease = true,
+				BundleAssemblies = true,
+				AotAssemblies = true,
+			};
+			proj.SetProperty (KnownProperties.TargetFrameworkVersion, "v5.1");
+			proj.SetAndroidSupportedAbis (supportedAbis);
+			proj.SetProperty ("EnableLLVM", enableLLVM.ToString ());
+			using (var b = CreateApkBuilder (path)) {
+				if (!b.CrossCompilerAvailable (supportedAbis))
+					Assert.Ignore ("Cross compiler was not available");
+				if (!b.GetSupportedRuntimes ().Any (x => supportedAbis == x.Abi))
+					Assert.Ignore ($"Runtime for {supportedAbis} was not available.");
+				b.ThrowOnBuildFailure = false;
+				Assert.AreEqual (expectedResult, b.Build (proj), "Build should have {0}.", expectedResult ? "succeeded" : "failed");
+				if (!expectedResult)
+					return;
+				foreach (var abi in supportedAbis.Split (new char [] { ';' })) {
+					var libapp = Path.Combine (Root, b.ProjectDirectory, proj.IntermediateOutputPath,
+						"bundles", abi, "libmonodroid_bundle_app.so");
+					Assert.IsTrue (File.Exists (libapp), abi + " libmonodroid_bundle_app.so does not exist");
+					var assemblies = Path.Combine (Root, b.ProjectDirectory, proj.IntermediateOutputPath,
+						"aot", abi, "libaot-UnnamedProject.dll.so");
+					Assert.IsTrue (File.Exists (assemblies), "{0} libaot-UnnamedProject.dll.so does not exist", abi);
+					var apk = Path.Combine (Root, b.ProjectDirectory,
+						proj.IntermediateOutputPath, "android", "bin", "UnnamedProject.UnnamedProject.apk");
+					using (var zipFile = ZipHelper.OpenZip (apk)) {
+						Assert.IsNotNull (ZipHelper.ReadFileFromZip (zipFile,
+							string.Format ("lib/{0}/libaot-UnnamedProject.dll.so", abi)),
+							"lib/{0}/libaot-UnnamedProject.dll.so should be in the UnnamedProject.UnnamedProject.apk", abi);
+						Assert.IsNull (ZipHelper.ReadFileFromZip (zipFile,
+							"assemblies/UnnamedProject.dll"),
+							"UnnamedProject.dll should not be in the UnnamedProject.UnnamedProject.apk");
+					}
+				}
+				Assert.AreEqual (expectedResult, b.Build (proj), "Second Build should have {0}.", expectedResult ? "succeeded" : "failed");
+				Assert.IsTrue (
+					b.Output.IsTargetSkipped ("_CompileJava"),
+					"the _CompileJava target should be skipped");
+				Assert.IsTrue (
+					b.Output.IsTargetSkipped ("_BuildApkEmbed"),
+					"the _BuildApkEmbed target should be skipped");
+			}
+		}
+
+
+		[Test]
+		[NonParallelizable]
+		[Category ("SmokeTests")]
+		public void BuildAMassiveApp ()
+		{
+			var testPath = Path.Combine ("temp", "BuildAMassiveApp");
+			TestOutputDirectories [TestContext.CurrentContext.Test.ID] = Path.Combine (Root, testPath);
+			var sb = new SolutionBuilder ("BuildAMassiveApp.sln") {
+				SolutionPath = Path.Combine (Root, testPath),
+				Verbosity = LoggerVerbosity.Diagnostic,
+			};
+			var app1 = new XamarinAndroidApplicationProject () {
+				TargetFrameworkVersion = sb.LatestTargetFrameworkVersion (),
+				ProjectName = "App1",
+				AotAssemblies = true,
+				IsRelease = true,
+				PackageReferences = {
+					KnownPackages.SupportDesign_27_0_2_1,
+					KnownPackages.SupportCompat_27_0_2_1,
+					KnownPackages.SupportCoreUI_27_0_2_1,
+					KnownPackages.SupportCoreUtils_27_0_2_1,
+					KnownPackages.SupportFragment_27_0_2_1,
+					KnownPackages.SupportMediaCompat_27_0_2_1,
+					KnownPackages.GooglePlayServicesMaps_42_1021_1,
+					KnownPackages.Xamarin_Build_Download_0_4_11,
+				},
+			};
+			//NOTE: BuildingInsideVisualStudio prevents the projects from being built as dependencies
+			sb.BuildingInsideVisualStudio = false;
+			app1.Imports.Add (new Import ("foo.targets") {
+				TextContent = () => @"<?xml version=""1.0"" encoding=""utf-16""?>
+<Project ToolsVersion=""4.0"" xmlns=""http://schemas.microsoft.com/developer/msbuild/2003"">
+<Target Name=""_CheckAbis"" BeforeTargets=""_DefineBuildTargetAbis"">
+	<PropertyGroup>
+		<AndroidSupportedAbis>armeabi-v7a;x86</AndroidSupportedAbis>
+		<AndroidSupportedAbis Condition=""Exists('$(MSBuildThisFileDirectory)..\..\..\..\Debug\lib\xamarin.android\xbuild\Xamarin\Android\lib\arm64-v8a\libmono-android.release.so')"">$(AndroidSupportedAbis);arm64-v8a</AndroidSupportedAbis>
+		<AndroidSupportedAbis Condition=""Exists('$(MSBuildThisFileDirectory)..\..\..\..\Debug\lib\xamarin.android\xbuild\Xamarin\Android\lib\x86_64\libmono-android.release.so')"">$(AndroidSupportedAbis);x86_64</AndroidSupportedAbis>
+	</PropertyGroup>
+	<Message Text=""$(AndroidSupportedAbis)"" />
+</Target>
+<Target Name=""_Foo"" AfterTargets=""_SetLatestTargetFrameworkVersion"">
+	<PropertyGroup>
+		<AotAssemblies Condition=""!Exists('$(MonoAndroidBinDirectory)" + Path.DirectorySeparatorChar + @"cross-arm')"">False</AotAssemblies>
+	</PropertyGroup>
+	<Message Text=""$(AotAssemblies)"" />
+</Target>
+</Project>
+",
+			});
+			app1.SetProperty (KnownProperties.AndroidUseSharedRuntime, "False");
+			sb.Projects.Add (app1);
+			var code = new StringBuilder ();
+			code.AppendLine ("using System;");
+			code.AppendLine ("namespace App1 {");
+			code.AppendLine ("\tpublic class AppCode {");
+			code.AppendLine ("\t\tpublic void Foo () {");
+			for (int i = 0; i < 128; i++) {
+				var libName = $"Lib{i}";
+				var lib = new XamarinAndroidLibraryProject () {
+					TargetFrameworkVersion = sb.LatestTargetFrameworkVersion (),
+					ProjectName = libName,
+					IsRelease = true,
+					OtherBuildItems = {
+						new AndroidItem.AndroidAsset ($"Assets\\{libName}.txt") {
+							TextContent = () => "Asset1",
+							Encoding = Encoding.ASCII,
+						},
+						new AndroidItem.AndroidAsset ($"Assets\\subfolder\\{libName}.txt") {
+							TextContent = () => "Asset2",
+							Encoding = Encoding.ASCII,
+						},
+					},
+					Sources = {
+						new BuildItem.Source ($"{libName}.cs") {
+							TextContent = () => @"using System;
+
+namespace "+ libName + @" {
+
+	public class " + libName + @" {
+		public static void Foo () {
+		}
+	}
+}"
+						},
+					}
+				};
+				var strings = lib.AndroidResources.First (x => x.Include () == "Resources\\values\\Strings.xml");
+				strings.TextContent = () => @"<?xml version=""1.0"" encoding=""utf-8""?>
+<resources>
+	<string name=""" + libName + @"_name"">" + libName + @"</string>
+</resources>";
+				sb.Projects.Add (lib);
+				app1.References.Add (new BuildItem.ProjectReference ($"..\\{libName}\\{libName}.csproj", libName, lib.ProjectGuid));
+				code.AppendLine ($"\t\t\t{libName}.{libName}.Foo ();");
+			}
+			code.AppendLine ("\t\t}");
+			code.AppendLine ("\t}");
+			code.AppendLine ("}");
+			app1.Sources.Add (new BuildItem.Source ("Code.cs") {
+				TextContent = () => code.ToString (),
+			});
+			Assert.IsTrue (sb.Build (new string [] { "Configuration=Release" }), "Solution should have built.");
+			Assert.IsTrue (sb.BuildProject (app1, "SignAndroidPackage"), "Build of project should have succeeded");
+			sb.Dispose ();
+		}
+
+		[Test]
+		public void HybridAOT ()
+		{
+			var proj = new XamarinAndroidApplicationProject () {
+				IsRelease = true,
+				AotAssemblies = true,
+			};
+			proj.SetProperty ("AndroidAotMode", "Hybrid");
+			// So we can use Mono.Cecil to open assemblies directly
+			proj.SetProperty ("AndroidEnableAssemblyCompression", "False");
+
+			using (var b = CreateApkBuilder ()) {
+				b.Build (proj);
+
+				var apk = Path.Combine (Root, b.ProjectDirectory, proj.OutputPath, $"{proj.PackageName}.apk");
+				FileAssert.Exists (apk);
+				using (var zip = ZipHelper.OpenZip (apk)) {
+					var entry = zip.ReadEntry ($"assemblies/{proj.ProjectName}.dll");
+					Assert.IsNotNull (entry, $"{proj.ProjectName}.dll should exist in apk!");
+					using (var stream = new MemoryStream ()) {
+						entry.Extract (stream);
+						stream.Position = 0;
+						using (var assembly = AssemblyDefinition.ReadAssembly (stream)) {
+							var type = assembly.MainModule.GetType ($"{proj.ProjectName}.MainActivity");
+							var method = type.Methods.First (m => m.Name == "OnCreate");
+							Assert.LessOrEqual (method.Body.Instructions.Count, 1, "OnCreate should have stripped method bodies!");
+						}
+					}
+				}
+			}
+		}
+
+		[Test]
+		public void NoSymbolsArgShouldReduceAppSize ([Values (true, false)] bool enableHybridAot)
+		{
+			var proj = new XamarinAndroidApplicationProject () {
+				IsRelease = true,
+				AotAssemblies = true,
+			};
+			var supportedAbi = "arm64-v8a";
+			proj.SetAndroidSupportedAbis (supportedAbi);
+			proj.SetProperty ("EnableLLVM", true.ToString ());
+			if (enableHybridAot)
+				proj.SetProperty ("AndroidAotMode", "Hybrid");
+
+			var xaAssemblySize = 0;
+			var xaAssemblySizeNoSymbol = 0;
+
+			using (var b = CreateApkBuilder ()) {
+				b.ThrowOnBuildFailure = false;
+				Assert.IsTrue (b.Build (proj), "First build should have succeeded.");
+				var apkPath = Path.Combine (Root, b.ProjectDirectory, proj.OutputPath, $"{proj.PackageName}-Signed.apk");
+				FileAssert.Exists (apkPath);
+				using (var apk = ZipHelper.OpenZip (apkPath)) {
+					xaAssemblySize = ZipHelper.ReadFileFromZip (apk, $"lib/{supportedAbi}/libaot-Mono.Android.dll.so").Length;
+				}
+				proj.SetProperty ("AndroidAotAdditionalArguments", "no-write-symbols");
+				Assert.IsTrue (b.Build (proj), "Second build should have succeeded.");
+				FileAssert.Exists (apkPath);
+				using (var apk = ZipHelper.OpenZip (apkPath)) {
+					xaAssemblySizeNoSymbol = ZipHelper.ReadFileFromZip (apk, $"lib/{supportedAbi}/libaot-Mono.Android.dll.so").Length;
+				}
+				Assert.IsTrue (xaAssemblySize > 0 && xaAssemblySizeNoSymbol > 0, $"Mono.Android.dll.so size was not updated after first or second build. Before: '{xaAssemblySize}' After: '{xaAssemblySizeNoSymbol}'.");
+				Assert.Less (xaAssemblySizeNoSymbol, xaAssemblySize, "Mono.Android.dll.so should have been smaller after 'no-write-symbols' build.");
+			}
+		}
+
+	}
+}
